[MCUX-77196][eqdc]Fix MISRA C 2012 violations
This patch fixes violations of MISRA C-2012 rules 10.1, 10.3, and 10.8
in the EQDC driver.

Changes include:
- Added explicit uint16_t casts for mask operations in EQDC_Init to
  ensure operands have appropriate essential type before bitwise OR
  operations.
- Changed conditional expressions from implicit boolean evaluation to
  explicit enum comparisons for outputPulseMode and
  revolutionCountCondition to avoid treating enum as boolean.
- Updated filter configuration fields (filterPhaseA, filterPhaseB,
  filterIndPre, filterHomEna) from 0U to false to match their boolean
  type.
- Changed return type of EQDC_GetEqdcLdok from uint8_t to uint16_t to
  match the actual register width and mask size.

+++ drivers/eqdc/fsl_eqdc.h
@@ -666,7 +666,7 @@ static inline void EQDC_SetEqdcLdok(EQDC_Type *base)
  *
  * @param base EQDC peripheral base address.
  */
-static inline uint8_t EQDC_GetEqdcLdok(EQDC_Type *base)
+static inline uint16_t EQDC_GetEqdcLdok(EQDC_Type *base)
 {
     return base->CTRL & EQDC_CTRL_LDOK_MASK;
 }
